\name{VCFAssociationHelper-package}
\alias{VCFAssociationHelper-package}
\alias{VCFAssociationHelper}
\docType{package}
\title{
  
  VCF read package

}
\description{
  
	Functions for transferring VCF file to Plink file and SSD file, reading SSD file for association test.  

}

\author{
Maintainer: Zhangchen Zhao <zczhao@umich.edu>
}


